singularity: add singularityce fork (#24201)
Since the two packages share a common history, the installation
procedure has been factored into a common base class.
